[Fusilli] Add .clang-tidy (#2595)
Adds .clang-tidy file to help promote a consistent code style and catch
bugs. This should only affect local development; it doesn't make any
changes to CI.

This has been mostly forked from [IREE's
clang-tidy](https://github.com/iree-org/iree/blob/main/compiler/src/.clang-tidy)
with the addition of `bugprone`, `performance`, and
`readability-identifier-naming` checks.


### Changes:
- Adds `.clang-tidy` file.
- Enforces `kGlobalConstant` naming for global constant variables.
- Fixes names that are against the conventions.
- Disables clang-tidy checks for endl.
- Disables `-bugprone-macro-parentheses` since it conflicts with some of
the logging macros.
- Makes other small changes suggested by clang tidy (std::move, make
functions static, etc).

---------

+++ sharkfuser/include/fusilli/attributes/tensor_attributes.h
@@ -109,6 +109,7 @@
 #include <ranges>
 #include <string>
 #include <type_traits>
+#include <utility>
 #include <variant>
 #include <vector>
 
@@ -181,7 +182,8 @@ generateStrideOrderPreservingFormat(const std::vector<int64_t> inputStride,
   if (outputDimSize > inputStride.size()) {
     size_t start = strideOrder.size();
     strideOrder.resize(outputDimSize);
-    std::iota(strideOrder.begin() + start, strideOrder.end(), start);
+    std::iota(strideOrder.begin() + static_cast<int64_t>(start),
+              strideOrder.end(), start);
   }
   return strideOrder;
 }
